requirements.txt
pyqt==5.9.2
opencv==4.0.1
numpy==1.21.2
